Fixed overflow error
